A Labber driver to make Labber understand numpy array-like arbitrary pulses
A Labber driver package to make Labber understand arbitrary pulse sequences designed in numpy arrays. The driver was integrated in the default singleQubit_Pulse_Generator driver's interface.
As such, it allows the production of all the possible mixes of externally imported pulses and the already available ones in the default singleQubit_Pulse_Generator driver.